+++ Project/Models/DTOs/UserMovie(Review)DTOs/UserMovie(Review)Update.cs
@@ -4,6 +4,6 @@
     {
      
         public Guid Id { get; set; }
-        public int? UserScore { get; set; }
+        public float? UserScore { get; set; }
     }
 }
